+++ src/components/SearchInputField.js
@@ -1,9 +1,19 @@
+/* eslint-disable react/jsx-props-no-spreading */
 import React from 'react';
+import PropTypes from 'prop-types';
 import InputField from './InputField';
 import SearchIcon from './SearchIcon';
 
-function SearchInputField(props) {
-    return <InputField adornment={SearchIcon()} {...props} />;
+function SearchInputField({ autoFocus, ...rest }) {
+    return <InputField adornment={SearchIcon()} autoFocus={autoFocus} {...rest} />;
 }
+
+SearchInputField.propTypes = {
+    autoFocus: PropTypes.bool
+};
+
+SearchInputField.defaultProps = {
+    autoFocus: false
+};
 
 export default SearchInputField;